domainを設計通りに作った

+++ back/src/domain/notification.go
@@ -1,7 +1,7 @@
 package domain
 
 type Notification struct {
-	WebhookId   uint64
+	WebhookId   uint32
 	RoomId      string
 	ServiceName string
 	Url         string
